+++ source/rafcon/core/states/execution_state.py
@@ -65,17 +65,15 @@ class ExecutionState(State):
         return str(self) == str(other) and self.script_text == other.script_text
 
     def __copy__(self):
-        input_data_ports = {elem_id: copy(elem) for elem_id, elem in self._input_data_ports.items()}
-        output_data_ports = {elem_id: copy(elem) for elem_id, elem in self._output_data_ports.items()}
+        input_data_ports = {key: copy(self._input_data_ports[key]) for key in self._input_data_ports.keys()}
+        output_data_ports = {key: copy(self._output_data_ports[key]) for key in self._output_data_ports.keys()}
         income = copy(self._income)
-        outcomes = {elem_id: copy(elem) for elem_id, elem in list(self._outcomes.items())}
+        outcomes = {elem_id: copy(self._outcomes[elem_id]) for elem_id in self._outcomes.keys()}
         state = self.__class__(self.name, self.state_id, input_data_ports, output_data_ports, income, outcomes, None,
                                safe_init=False)
-        try:
-            # use setter here! the acutal value, which is changed is self._script.script!
-            state.script_text = deepcopy(self.script_text)
-        except:
-            pass  # Tolerate script compilation errors
+
+        state.script_text = deepcopy(self.script_text)
+
         state._description = deepcopy(self.description)
         state._semantic_data = deepcopy(self.semantic_data)
         state._file_system_path = self.file_system_path