+++ source/rafcon/gui/controllers/modification_history.py
@@ -175,7 +175,7 @@ class ModificationHistoryTreeController(ExtendedController):
                 self.update(None, None, None)
         self.no_cursor_observation = False
 
-    def undo(self, key_value, modifier_mask):
+    def undo(self, key_value, modifier_mask, **kwargs):
         """Undo for selected state-machine if no state-source-editor is open and focused in states-editor-controller.
 
         :return: True if a undo was performed, False if focus on source-editor.
@@ -196,7 +196,7 @@ class ModificationHistoryTreeController(ExtendedController):
         else:
             logger.debug("Undo is not possible now as long as no state_machine is selected.")
 
-    def redo(self, key_value, modifier_mask):
+    def redo(self, key_value, modifier_mask, **kwargs):
         """Redo for selected state-machine if no state-source-editor is open and focused in states-editor-controller.
 
         :return: True if a redo was performed, False if focus on source-editor.