+++ source/rafcon/gui/controllers/right_click_menu/state.py
@@ -45,6 +45,13 @@ logger = log.get_logger(__name__)
 
 class StateMachineRightClickMenu(object):
 
+    """ A class for handling all right click menus inside rafcon.
+    Examples are the right click menu of the state machine tree or the gaphas editor.
+
+    """
+
+    menu_position = None
+
     def __init__(self, state_machine_manager_model=None):
         if state_machine_manager_model is None:
             from rafcon.gui.singleton import state_machine_manager_model
@@ -278,16 +285,16 @@ class StateMachineRightClickMenu(object):
         self.shortcut_manager.trigger_action("is_start_state", None, None)
 
     def on_add_execution_state_activate(self, widget=None, data=None):
-        self.shortcut_manager.trigger_action('add_execution_state', None, None)
+        self.shortcut_manager.trigger_action('add_execution_state', None, None, cursor_position=self.menu_position)
 
     def on_add_hierarchy_state_activate(self, widget=None, data=None):
-        self.shortcut_manager.trigger_action('add_hierarchy_state', None, None)
+        self.shortcut_manager.trigger_action('add_hierarchy_state', None, None, cursor_position=self.menu_position)
 
     def on_add_preemptive_state_activate(self, widget=None, data=None):
-        self.shortcut_manager.trigger_action('add_preemptive_state', None, None)
+        self.shortcut_manager.trigger_action('add_preemptive_state', None, None, cursor_position=self.menu_position)
 
     def on_add_barrier_state_activate(self, widget=None, data=None):
-        self.shortcut_manager.trigger_action('add_barrier_state', None, None)
+        self.shortcut_manager.trigger_action('add_barrier_state', None, None, cursor_position=self.menu_position)
 
     def on_add_outcome(self, widget=None, data=None):
         self.shortcut_manager.trigger_action('add_outcome', None, None)
@@ -444,12 +451,19 @@ class StateRightClickMenuControllerOpenGLEditor(StateMachineRightClickMenuContro
 
 
 class StateRightClickMenuGaphas(StateMachineRightClickMenu):
+    """ A class for handling the right click menu inside gaphas
+
+    """
 
     def activate_menu(self, event, menu):
         # logger.info("activate_menu by " + self.__class__.__name__)
         selection = gui_singletons.state_machine_manager_model.get_selected_state_machine_model().selection
         if len(selection.states) > 0 or len(selection.scoped_variables) > 0:
+            from rafcon.gui.helpers.coordinates import screen2main_window
             menu.popup(None, None, None, None, event.get_button()[1], event.time)
+            # The pointer in screen coordinates.
+            pointer = menu.get_root_window().get_pointer()
+            self.menu_position = screen2main_window((pointer.x, pointer.y))
             return True
         else:
             return False
@@ -460,7 +474,7 @@ class StateRightClickMenuGaphas(StateMachineRightClickMenu):
 
     def on_paste_activate(self, widget, data=None):
         # logger.info("trigger gaphas paste")
-        self.shortcut_manager.trigger_action("paste", None, None)
+        self.shortcut_manager.trigger_action("paste", None, None, cursor_position=self.menu_position)
 
     def on_cut_activate(self, widget, data=None):
         # logger.info("trigger gaphas cut")